+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The changelog format follows these rules: [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and sticks to [semantic versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - UNRELEASED
+### Added
+- Time journals can now also be analysed.
+
 ## [0.0.3] - 2018-06-08
 ### Added
 - It is now possible to add additional accounts with amount, which are not in the ledger journal.